eliminates prototypes.h (implements #921, #868)
