+++ src/braces.cpp
@@ -8,7 +8,6 @@
 #include "braces.h"
 #include "uncrustify_types.h"
 #include "chunk_list.h"
-#include "prototypes.h"
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
