+++ src/chunk_list.h
@@ -44,33 +44,143 @@ chunk_t *chunk_get_tail(void);
 chunk_t *chunk_get_next(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
 chunk_t *chunk_get_prev(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
 
+
+/**
+ * Swaps the two chunks.
+ *
+ * @param pc1  The first chunk
+ * @param pc2  The second chunk
+ */
 void chunk_swap(chunk_t *pc1, chunk_t *pc2);
+
+
+/**
+ * Swaps two lines that are started with the specified chunks.
+ *
+ * @param pc1  The first chunk of line 1
+ * @param pc2  The first chunk of line 2
+ */
 void chunk_swap_lines(chunk_t *pc1, chunk_t *pc2);
+
+
+/**
+ * Finds the first chunk on the line that pc is on.
+ * This just backs up until a newline or NULL is hit.
+ *
+ * given: [ a - b - c - n1 - d - e - n2 ]
+ * input: [ a | b | c | n1 ] => a
+ * input: [ d | e | n2 ]     => d
+ */
 chunk_t *chunk_first_on_line(chunk_t *pc);
 
+
+/**
+ * Gets the next NEWLINE chunk
+ */
 chunk_t *chunk_get_next_nl(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
+
+
 chunk_t *chunk_get_next_nc(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
+
+
+/**
+ * Gets the next non-NEWLINE chunk
+ */
 chunk_t *chunk_get_next_nnl(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
+
+
+/**
+ * Gets the next non-NEWLINE and non-comment chunk
+ */
 chunk_t *chunk_get_next_ncnl(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
+
+
+/**
+ * Gets the next non-NEWLINE and non-comment chunk, non-preprocessor chunk
+ */
 chunk_t *chunk_get_next_ncnlnp(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
+
+
+/**
+ * Gets the next chunk not in or part of balanced square
+ * brackets. This handles stacked [] instances to accommodate
+ * multi-dimensional array declarations
+ *
+ * @param cur     Starting chunk
+ * @return        NULL or the next chunk not in or part of
+ *                square brackets
+ */
 chunk_t *chunk_get_next_nisq(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
+
 
 chunk_t *chunk_get_next_nblank(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
 chunk_t *chunk_get_prev_nblank(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
 
+
+/**
+ * Gets the prev NEWLINE chunk
+ */
 chunk_t *chunk_get_prev_nl(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
+
+
+/**
+ * Gets the prev non-comment chunk
+ */
 chunk_t *chunk_get_prev_nc(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
+
+
+/**
+ * Gets the prev non-NEWLINE chunk
+ */
 chunk_t *chunk_get_prev_nnl(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
+
+
+/**
+ * Gets the prev non-NEWLINE and non-comment chunk
+ */
 chunk_t *chunk_get_prev_ncnl(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
+
+
 chunk_t *chunk_get_prev_ncnlnp(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
 
+
+/**
+ * Grabs the next chunk of the given type at the level.
+ *
+ * @param cur     Starting chunk
+ * @param type    The type to look for
+ * @param level   -1 or ANY_LEVEL (any level) or the level to match
+ * @return        NULL or the match
+ */
 chunk_t *chunk_get_next_type(chunk_t *cur, c_token_t type, int level, chunk_nav_t nav = CNAV_ALL);
+
+
+/**
+ * Grabs the prev chunk of the given type at the level.
+ *
+ * @param cur     Starting chunk
+ * @param type    The type to look for
+ * @param level   -1 or ANY_LEVEL (any level) or the level to match
+ * @return        NULL or the match
+ */
 chunk_t *chunk_get_prev_type(chunk_t *cur, c_token_t type, int level, chunk_nav_t nav = CNAV_ALL);
 
+
 chunk_t *chunk_get_next_str(chunk_t *cur, const char *str, size_t len, int level, chunk_nav_t nav = CNAV_ALL);
+
+
 chunk_t *chunk_get_prev_str(chunk_t *cur, const char *str, size_t len, int level, chunk_nav_t nav = CNAV_ALL);
 
+
+/**
+ * Gets the next non-vbrace chunk
+ */
 chunk_t *chunk_get_next_nvb(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
+
+
+/**
+ * Gets the prev non-vbrace chunk
+ */
 chunk_t *chunk_get_prev_nvb(chunk_t *cur, chunk_nav_t nav = CNAV_ALL);
 
 
@@ -273,6 +383,9 @@ bool chunk_is_ptr_operator(chunk_t *pc)
 }
 
 
+/**
+ * Check to see if there is a newline between the two chunks
+ */
 bool chunk_is_newline_between(chunk_t *start, chunk_t *end);
 
 
@@ -380,7 +493,10 @@ bool chunk_is_forin(chunk_t *pc)
 
 
 void set_chunk_type_real(chunk_t *pc, c_token_t tt);
+
+
 void set_chunk_parent_real(chunk_t *pc, c_token_t tt);
+
 
 #define set_chunk_type(pc, tt)      do { \
       LOG_FUNC_CALL();                   \
@@ -392,7 +508,9 @@ void set_chunk_parent_real(chunk_t *pc, c_token_t tt);
       set_chunk_parent_real((pc), (tt)); \
 } while (false)
 
+
 void chunk_flags_set_real(chunk_t *pc, UINT64 clr_bits, UINT64 set_bits);
+
 
 #define chunk_flags_upd(pc, cc, ss)    do {   \
       LOG_FUNC_CALL();                        \
