+++ src/combine.h
@@ -51,6 +51,29 @@ void flag_series(chunk_t *start, chunk_t *end, UINT64 set_flags, UINT64 clr_flag
 
 
 /**
+ * Sets the parent of the open paren/brace/square/angle and the closing.
+ * Note - it is assumed that pc really does point to an open item and the
+ * close must be open + 1.
+ *
+ * @param start   The open paren
+ * @param parent  The type to assign as the parent
+ * @return        The chunk after the close paren
+ */
+chunk_t *set_paren_parent(chunk_t *start, c_token_t parent);
+
+
+/**
+ * This is called on every chunk.
+ * First on all non-preprocessor chunks and then on each preprocessor chunk.
+ * It does all the detection and classifying.
+ * This is only called by fix_symbols.
+ * The three parameters never get the value NULL.
+ * it is not necessary to test.
+ */
+void do_symbol_check(chunk_t *prev, chunk_t *pc, chunk_t *next);
+
+
+/**
  * Skips over the rest of the template if ang_open is indeed a CT_ANGLE_OPEN.
  * Points to the chunk after the CT_ANGLE_CLOSE.
  * If the chunk isn't an CT_ANGLE_OPEN, then it is returned.
