+++ src/md5.cpp
@@ -158,10 +158,6 @@ void MD5::Update(const void *data, UINT32 len)
 } // MD5::Update
 
 
-/**
- * Final wrapup - pad to 64-byte boundary with the bit pattern
- * 1 0* (64-bit count of bits processed, MSB-first)
- */
 void MD5::Final(UINT8 digest[16])
 {
    /* Compute number of bytes mod 64 */
