+++ src/newlines.cpp
@@ -20,7 +20,6 @@
 #include "newlines.h"
 #include "uncrustify_types.h"
 #include "chunk_list.h"
-#include "prototypes.h"
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
@@ -414,10 +413,6 @@ static void setup_newline_add(chunk_t *prev, chunk_t *nl, chunk_t *next)
 }
 
 
-/**
- * Add a newline before the chunk if there isn't already a newline present.
- * Virtual braces are skipped, as they do not contribute to the output.
- */
 chunk_t *newline_add_before(chunk_t *pc)
 {
    LOG_FUNC_ENTRY();
@@ -457,10 +452,6 @@ chunk_t *newline_force_before(chunk_t *pc)
 }
 
 
-/**
- * Add a newline after the chunk if there isn't already a newline present.
- * Virtual braces are skipped, as they do not contribute to the output.
- */
 chunk_t *newline_add_after(chunk_t *pc)
 {
    LOG_FUNC_ENTRY();
@@ -582,21 +573,6 @@ static void newline_min_after(chunk_t *ref, size_t count, UINT64 flag)
 } // newline_min_after
 
 
-/**
- * Add a newline between two tokens.
- * If there is already a newline between then, nothing is done.
- * Otherwise a newline is inserted.
- *
- * If end is CT_BRACE_OPEN and a comment and newline follow, then
- * the brace open is moved instead of inserting a newline.
- *
- * In this situation:
- *    if (...) { //comment
- *
- * you get:
- *    if (...)   //comment
- *    {
- */
 chunk_t *newline_add_between(chunk_t *start, chunk_t *end)
 {
    LOG_FUNC_ENTRY();
@@ -649,16 +625,6 @@ chunk_t *newline_add_between(chunk_t *start, chunk_t *end)
 } // newline_add_between
 
 
-/**
- * Removes any CT_NEWLINE or CT_NL_CONT between start and end.
- * Start must be before end on the chunk list.
- * If the 'PCF_IN_PREPROC' status differs between two tags, we can't remove
- * the newline.
- *
- * @param start   The starting chunk (if it is a newline, it will be removed!)
- * @param end     The ending chunk (will not be removed, even if it is a newline)
- * @return        true/false - removed something
- */
 void newline_del_between(chunk_t *start, chunk_t *end)
 {
    LOG_FUNC_ENTRY();
@@ -1881,12 +1847,6 @@ static void newline_iarf_pair(chunk_t *before, chunk_t *after, argval_t av)
 }
 
 
-/**
- * Does a simple Ignore, Add, Remove, or Force after the given chunk
- *
- * @param pc   The chunk
- * @param av   The IARF value
- */
 void newline_iarf(chunk_t *pc, argval_t av)
 {
    LOG_FUNC_ENTRY();
@@ -2301,10 +2261,6 @@ static bool one_liner_nl_ok(chunk_t *pc)
 } // one_liner_nl_ok
 
 
-/**
- * Clears the PCF_ONE_LINER flag on the current line.
- * Done right before inserting a newline.
- */
 void undo_one_liner(chunk_t *pc)
 {
    LOG_FUNC_ENTRY();
@@ -2376,9 +2332,6 @@ static void nl_create_one_liner(chunk_t *vbrace_open)
 }
 
 
-/**
- * Remove all extra newlines.
- */
 void newlines_remove_newlines(void)
 {
    LOG_FUNC_ENTRY();
@@ -2414,9 +2367,6 @@ void newlines_remove_newlines(void)
 }
 
 
-/**
- * Step through all chunks.
- */
 void newlines_cleanup_braces(bool first)
 {
    LOG_FUNC_ENTRY();
@@ -3096,9 +3046,6 @@ void newline_after_multiline_comment(void)
 }
 
 
-/**
- * Handle insertion of blank lines after label colons
- */
 void newline_after_label_colon(void)
 {
    LOG_FUNC_ENTRY();
@@ -3115,9 +3062,6 @@ void newline_after_label_colon(void)
 }
 
 
-/**
- * Handle insertion/removal of blank lines before if/for/while/do and functions
- */
 void newlines_insert_blank_lines(void)
 {
    LOG_FUNC_ENTRY();
@@ -3169,10 +3113,6 @@ void newlines_insert_blank_lines(void)
 } // newlines_insert_blank_lines
 
 
-/**
- * Handle removal of extra blank lines in functions
- * x <= 0: do nothing, x > 0: allow max x-1 blank lines
- */
 void newlines_functions_remove_extra_blank_lines(void)
 {
    LOG_FUNC_ENTRY();
@@ -3238,8 +3178,6 @@ void newlines_squeeze_ifdef(void)
                pnl = chunk_get_prev_nl(pc);
             }
 
-            chunk_t *tmp1;
-            chunk_t *tmp2;
             if (nnl != NULL)
             {
                if (pnl != NULL)
@@ -3250,8 +3188,8 @@ void newlines_squeeze_ifdef(void)
                      pnl->nl_count = 1;
                      MARK_CHANGE();
 
-                     tmp1 = chunk_get_prev_nnl(pnl);
-                     tmp2 = chunk_get_prev_nnl(nnl);
+                     chunk_t *tmp1 = chunk_get_prev_nnl(pnl);
+                     chunk_t *tmp2 = chunk_get_prev_nnl(nnl);
 
                      LOG_FMT(LNEWLINE, "%s: moved from after line %zu to after %zu\n",
                              __func__, tmp1->orig_line, tmp2->orig_line);
@@ -3262,7 +3200,7 @@ void newlines_squeeze_ifdef(void)
                {
                   if (nnl->nl_count > 1)
                   {
-                     tmp1 = chunk_get_prev_nnl(nnl);
+                     chunk_t *tmp1 = chunk_get_prev_nnl(nnl);
                      LOG_FMT(LNEWLINE, "%s: trimmed newlines after line %zu from %zu\n",
                              __func__, tmp1->orig_line, nnl->nl_count);
                      nnl->nl_count = 1;
@@ -3366,12 +3304,6 @@ void newlines_eat_start_end(void)
 } // newlines_eat_start_end
 
 
-/**
- * Searches for a chunk of type chunk_type and moves them, if needed.
- * Will not move tokens that are on their own line or have other than
- * exactly 1 newline before (UO_pos_comma == TRAIL) or after (UO_pos_comma == LEAD).
- * We can't remove a newline if it is right before a preprocessor.
- */
 void newlines_chunk_pos(c_token_t chunk_type, tokenpos_e mode)
 {
    LOG_FUNC_ENTRY();
@@ -3519,10 +3451,6 @@ void newlines_chunk_pos(c_token_t chunk_type, tokenpos_e mode)
 } // newlines_chunk_pos
 
 
-/**
- * Searches for CT_CLASS_COLON and moves them, if needed.
- * Also breaks up the args
- */
 void newlines_class_colon_pos(c_token_t tok)
 {
    LOG_FUNC_ENTRY();
@@ -3692,13 +3620,6 @@ static void _blank_line_max(chunk_t *pc, const char *text, uncrustify_options uo
 #define blank_line_max(pc, op)    _blank_line_max(pc, # op, op)
 
 
-/**
- * Scans for newline tokens and changes the nl_count.
- * A newline token has a minimum nl_count of 1.
- * Note that a blank line is actually 2 newlines, unless the newline is the
- * first chunk.
- * So, most comparisons have +1 below.
- */
 void do_blank_lines(void)
 {
    LOG_FUNC_ENTRY();
