+++ src/output.h
@@ -37,4 +37,5 @@ void output_parsed(FILE *pfile);
  */
 void add_long_preprocessor_conditional_block_comment(void);
 
+
 #endif /* OUTPUT_H_INCLUDED */
