+++ src/parse_frame.cpp
@@ -7,7 +7,6 @@
  */
 #include "parse_frame.h"
 #include "uncrustify_types.h"
-#include "prototypes.h"
 #include "chunk_list.h"
 #include "uncrustify.h"
 #include <cstdio>
@@ -16,6 +15,12 @@
 
 
 static void pf_log_frms(log_sev_t logsev, const char *txt, struct parse_frame *pf);
+
+
+/**
+ * Logs the entire parse frame stack
+ */
+void pf_log_all(log_sev_t logsev);
 
 
 /**
@@ -60,9 +65,6 @@ static void pf_log_frms(log_sev_t logsev, const char *txt, parse_frame_t *pf)
 }
 
 
-/**
- * Logs the entire parse frame stack
- */
 void pf_log_all(log_sev_t logsev)
 {
    LOG_FMT(logsev, "##=- Parse Frame : %d entries\n", cpd.frame_count);
@@ -77,19 +79,12 @@ void pf_log_all(log_sev_t logsev)
 }
 
 
-/**
- * Copies src to dst.
- */
 void pf_copy(parse_frame_t *dst, const parse_frame_t *src)
 {
    memcpy(dst, src, sizeof(parse_frame_t));
 }
 
 
-/**
- * Push a copy of the parse frame onto the stack.
- * This is called on #if and #ifdef.
- */
 void pf_push(parse_frame_t *pf)
 {
    static int ref_no = 1;
@@ -105,11 +100,6 @@ void pf_push(parse_frame_t *pf)
 }
 
 
-/**
- * Push a copy of the parse frame onto the stack, under the tos.
- * If this were a linked list, just add before the last item.
- * This is called on the first #else and #elif.
- */
 void pf_push_under(parse_frame_t *pf)
 {
    LOG_FMT(LPF, "%s(%d): before count = %d\n", __func__, __LINE__, cpd.frame_count);
@@ -128,10 +118,6 @@ void pf_push_under(parse_frame_t *pf)
 }
 
 
-/**
- * Copy the top item off the stack into pf.
- * This is called on #else and #elif.
- */
 void pf_copy_tos(parse_frame_t *pf)
 {
    if (cpd.frame_count > 0)
@@ -152,9 +138,6 @@ static void pf_copy_2nd_tos(parse_frame_t *pf)
 }
 
 
-/**
- * Deletes the top frame from the stack.
- */
 void pf_trash_tos(void)
 {
    if (cpd.frame_count > 0)
@@ -165,10 +148,6 @@ void pf_trash_tos(void)
 }
 
 
-/**
- * Pop the top item off the stack and copy into pf.
- * This is called on #endif
- */
 void pf_pop(parse_frame_t *pf)
 {
    if (cpd.frame_count > 0)
@@ -180,9 +159,6 @@ void pf_pop(parse_frame_t *pf)
 }
 
 
-/**
- * Returns the pp_indent to use for this line
- */
 int pf_check(parse_frame_t *frm, chunk_t *pc)
 {
    int in_ifdef = frm->in_ifdef;
