+++ src/punctuators.cpp
@@ -1,22 +1,13 @@
 /**
  * @file punctuators.cpp
- * Manages the table of punctutators.
+ * Manages the table of punctuators.
  *
  * @author  Ben Gardner
  * @license GPL v2+
  */
+#include "punctuators.h"
 #include "uncrustify_types.h"
 #include "prototypes.h"
-
-
-/**
- *
- *   The file "punctuators.h" was generated by punc.py from this file.
- *   If you change one of the below symbol tables, re-run punc.py.
- *   $ python scripts/punc.py > src/punctuators.h
- *
- *   NOTE: the tables below do not need to be sorted.
- */
 
 
 /* 4-char symbols */
@@ -119,7 +110,8 @@ static const chunk_tag_t symbols1[] =
    { "~", CT_INV,          LANG_ALL                            },
 };
 
-#include "punctuators.h"
+/* keep this include after the symbols1 - symbols4 */
+#include "punc_table.h"
 
 
 const chunk_tag_t *find_punctuator(const char *str, int lang_flags)
