+++ src/semicolons.h
@@ -22,4 +22,5 @@
  */
 void remove_extra_semicolons(void);
 
+
 #endif /* SEMICOLONS_H_INCLUDED */
