+++ src/sorting.h
@@ -15,4 +15,5 @@
 
 void sort_imports(void);
 
+
 #endif /* SORTING_H_INCLUDED */
