+++ src/tokenize.h
@@ -11,6 +11,18 @@
 #include "uncrustify_types.h"
 
 
+/**
+ * This function parses or tokenizes the whole buffer into a list.
+ * It has to do some tricks to parse preprocessors.
+ *
+ * If output_text() were called immediately after, two things would happen:
+ *  - trailing whitespace are removed.
+ *  - leading space & tabs are converted to the appropriate format.
+ *
+ * All the tokens are inserted before ref. If ref is NULL, they are inserted
+ * at the end of the list.  Line numbers are relative to the start of the data.
+ */
 void tokenize(const deque<int> &data, chunk_t *ref);
+
 
 #endif /* TOKENIZE_H_INCLUDED */
