+++ src/tokenize_cleanup.h
@@ -12,6 +12,8 @@
 
 
 void tokenize_cleanup(void);
+
+
 void split_off_angle_close(chunk_t *pc);
 
 #endif /* TOKENIZ_CLEANUP_H_INCLUDED */
