+++ src/unicode.cpp
@@ -8,7 +8,6 @@
 #include "unicode.h"
 #include "uncrustify.h"
 #include "uncrustify_types.h"
-#include "prototypes.h"
 #include "unc_ctype.h"
 #include <cstring>
 #include <cstdlib>
