+++ src/width.h
@@ -11,6 +11,10 @@
 #include "uncrustify_types.h"
 
 
+/**
+ * Step forward until a token goes beyond the limit and then call split_line()
+ * to split the line at or before that point.
+ */
 void do_code_width(void);
 
 #endif /* WIDTH_H_INCLUDED */
